dr_pcx.h: drpcx__decode_8bit(): 64-bit fix
position is implemented as a pointer...

+++ ttbasic/dr_pcx.h
@@ -487,7 +487,7 @@ dr_bool32 drpcx__decode_8bit(drpcx* pPCX)
             if (!palette256)
               return DR_FALSE;
 
-            uint32_t position = pPCX->onRead(pPCX->pUserData, (void *)-1, 0);
+            unsigned long position = pPCX->onRead(pPCX->pUserData, (void *)-1, 0);
 
             for (dr_uint32 y = 0; y < pPCX->height; ++y)
             {
